Move GraphqlController to under Api namespace
